+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to Strange Matter will be documented in this file.
 
 ### Fixed
 - **Warp Gate Entity Language Entry**: Added missing language entry for Warp Gate Entity
+- **Hoverboard**: Fixed issue with hoverboard sinking inside slabs and causing fall damage. Smoothed out turning as well.
+- **Improved slider responsiveness**: Improved the responsiveness of the gravity minigames slider.
 
 ## [0.2.5] - 2025-10-20